backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.8.0
pydantic>=2.11.0
python-dotenv>=1.0.1
